LayoutTests/media/media-can-play-av1.html

<!DOCTYPE html>
<html>
    <head>
        <script src="video-test.js"></script>
        <script>
         function start() {
             video = mediaElement = document.getElementsByTagName('video')[0];

             testExpected("video.canPlayType('video/mp4; codecs=av1')", "");
             testExpected("video.canPlayType('video/webm; codecs=av1')", "");

             testExpected("video.canPlayType('video/mp4; codecs=av01.0.04M.08')", "");
             testExpected("video.canPlayType('video/webm; codecs=av01.0.04M.08')", "");

             endTest();
         }
        </script>
    </head>
    <body onload="start()">
        <video controls></video>
        <p>Test HTMLMediaElement <em>canPlayType()</em> method for the AV1 codec.</p>
        <p>These tests may be expected to fail if the WebKit port does not support the format.</p>
    </body>
</html>
